#16 feat: alert->required로 변경
